Generic parse_generic() functions fix #32 and #34
Branching parse_generic() that do all common checks on the query and the
answer. And then parse either journals or publisher depending on the
detected pattern (it auto-determines which parser to use).
Had to modify the tests accordingly as the warning messages changed.
All higher-level functions now call 'parse_generic()'